move event binding func back into index now that its only called from one loc
